# apps/coupon-api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDADAN Coupon API",
    "description": "상점별 쿠폰·도장 발급 시스템 API. 모든 경로는 `/api/coupon/v1` 아래에 있으며, JSON 필드는 snake_case, 식별자는 UUID 문자열, 시각은 RFC 3339 UTC, 금액은 원 단위 정수입니다.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7810",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/coupon/v1/admin/privacy/erasures": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Erasure requests and their outcomes (ADMIN-006).",
        "operationId": "list_erasures",
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErasuresResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Queue an erasure (ADMIN-006 삭제).",
        "description": "The request does not erase anything: it registers the obligation, records the case that\nauthorises it, and queues a job that runs after the grace period. §17.3's hold check\nhappens both here and in the job, because a dispute can open in between.",
        "operationId": "request_erasure",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestErasureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErasureRecord"
                }
              }
            }
          },
          "403": {
            "description": "This role may not request erasure"
          },
          "404": {
            "description": "No such subject or case"
          },
          "422": {
            "description": "LEGAL_HOLD_ACTIVE"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/admin/privacy/erasures/reapply": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Re-apply every completed erasure (§18.5).",
        "description": "The runbook step after a restore. Safe to run at any time — the erasure is idempotent —\nand `reapplied > 0` outside a restore means something wrote an erased subject back,\nwhich is worth an alert rather than a shrug.",
        "operationId": "reapply_erasures",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "How many subjects had to be erased again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReapplyResult"
                }
              }
            }
          },
          "403": {
            "description": "This role may not replay the deletion ledger"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/admin/retention-policies": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "The configured retention periods (§17.3).",
        "operationId": "list_retention_policies",
        "responses": {
          "200": {
            "description": "Per-category retention",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPoliciesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/admin/retention-policies/{data_category}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change one category's retention period.",
        "description": "§17.3 makes these configuration rather than constants precisely so they can change\nwithout a deploy — but a shortened period destroys evidence, so it demands a fresh\nsign-in, a written legal basis and a reason, and it lands in the audit trail.",
        "operationId": "patch_retention_policy",
        "parameters": [
          {
            "name": "data_category",
            "in": "path",
            "description": "PROFILE, CONSENT, TRANSACTION, …",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRetentionPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicy"
                }
              }
            }
          },
          "403": {
            "description": "This role may not change retention"
          },
          "404": {
            "description": "No such category"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErasureRecord": {
        "type": "object",
        "required": [
          "id",
          "subject_user_id",
          "status",
          "requested_at",
          "execute_after",
          "deletion_scope",
          "applied_scopes",
          "reapply_count"
        ],
        "properties": {
          "applied_scopes": {},
          "blocked_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "case_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "deletion_scope": {},
          "execute_after": {
            "type": "string",
            "format": "date-time"
          },
          "executed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "job_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "pseudonym_label": {
            "type": [
              "string",
              "null"
            ]
          },
          "reapply_count": {
            "type": "integer",
            "format": "int32"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "subject_user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ErasuresResponse": {
        "type": "object",
        "required": [
          "erasures"
        ],
        "properties": {
          "erasures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErasureRecord"
            }
          }
        }
      },
...
      "ReapplyResult": {
        "type": "object",
        "description": "What replaying the ledger did (§18.5).",
        "required": [
          "examined",
          "reapplied"
        ],
        "properties": {
          "examined": {
            "type": "integer",
            "format": "int64"
          },
          "reapplied": {
            "type": "integer",
            "format": "int64",
            "description": "How many subjects were found alive again and re-erased. Non-zero after a restore is\nexpected; non-zero at any other time means something wrote a person back."
          }
        }
      },
...
      "RequestErasureRequest": {
        "type": "object",
        "description": "`POST /admin/privacy/erasures` (ADMIN-006 삭제).",
        "required": [
          "subject_user_id",
          "case_id",
          "reason"
        ],
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid",
            "description": "§3.3 and ADMIN-006: the case is the record of who asked, when, and by what deadline."
          },
          "reason": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Categories to erase. Empty means every category the policy table marks erasable."
          },
          "subject_user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RetentionPoliciesResponse": {
        "type": "object",
        "required": [
          "policies"
        ],
        "properties": {
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetentionPolicy"
            }
          }
        }
      },
      "RetentionPolicy": {
        "type": "object",
        "description": "One row of `retention_policies` (§17.3).",
        "required": [
          "data_category",
          "retention_days",
          "legal_basis",
          "erasable_on_request",
          "active",
          "updated_at"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "data_category": {
            "type": "string"
          },
          "erasable_on_request": {
            "type": "boolean",
            "description": "Whether a subject may have this category erased on request, or whether a statutory\nretention keeps it until the period runs out (§17.3, ADMIN-006 삭제)."
          },
          "legal_basis": {
            "type": "string"
          },
          "retention_days": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateRetentionPolicyRequest": {
        "type": "object",
        "description": "`PATCH /admin/retention-policies/:category`.",
        "required": [
          "retention_days",
          "legal_basis",
          "reason"
        ],
        "properties": {
          "erasable_on_request": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "legal_basis": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "description": "§11.5 requires a reason on every administrative change."
          },
          "retention_days": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token"
      }
    }
  },
...
}